agent_mgmt/requirements.txt
Django>=2.0
psycopg2
django-treebeard